requirements.txt
asgiref==3.7.2
Django==5.0
PyJWT==2.8.0
sqlparse==0.4.4
